OMAP4: PM: TWL603[02]: update min, max voltages
In TWL6030 depending on the value of SMPS_OFFSET efuse register the
voltage range supported in standard mode can be either between
~0.6V - ~1.3V or ~0.7V - ~1.4V. In TWL6030 ES1.0, SMPS_OFFSET efuse
is programmed to all 0's where as starting from TWL6030 ES1.1 the
efuse is programmed to 1.

Since the operational support is from TWL6030 ES1.1, and some of
the TWL6030 firmwares allow us to override the SMPS_OFFSET, we use
the 1.41V as the max voltage and min as 0V as the voltage values
from ~0.7 - ~1.4v are equidistant for each step value or in other
words, 1 step is symmetric 12.66..mV in this range.

In case we are unable to do anything about the offset, then we throw
out a warning and continue as the system should still be able to
function despite the higher value configuration.

Change-Id: I6f93b7ba4694d96f96ce53a1bbe8f058e62c4bc8
[nm@ti.com: rewrite, split up and new controls]
Signed-off-by: Nishanth Menon <nm@ti.com>
[oleksandr.kozaruk@ti.com: original patch]
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>

Conflicts:
	arch/arm/mach-omap2/omap_twl.c