ASoC: ABE: Increase volume ramps for DMic's
Most digital microphones need about 10 ms to settle after setting bias
and clocks.  The OMAP DMic module is not able to set the clocks before
beginning the transfer, creating a full-scale pop at the beginning of
the transfer.

This patch works around this behavior by increasing the volume ramp
from 2ms to 50ms for DMic's.

Change-Id: I55fe727fb624965ad22c0f9087c8b426b18080ef
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>