OMAP4: DPLL cascading: usage of DPLL cascading blocker API on HSI activity
When HSI driver is probed, it sets a constraint to keep HSI Fclk at 96 MHz.
But this constraint was not released, thus it was not possible to enter
OPPLP (used for DPLL cascading).

Now DPLL cascading blocker constraint is set and release based on HSI
activity.

[Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>:
Ported and adapted to p-android-omap-3.0]

Change-Id: Id8ace8f8a4f0a24641e51023346a745ecaaaf423
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>
Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>

Conflicts:
	drivers/omap_hsi/hsi_driver.c