Enable CONFIG_MODVERSIONS
CONFIG_MODVERSIONS allows modules to be used
with a different kernel than what it was built
against. When CONFIG_MODVERSIONS is enabled,
the kernel will add a special piece of versioning
info to every exported symbol. This versioning info
is calculated by a CRC algorithm which recursively
looks at the symbols associated structs, unions,
enums, and typedefs. When an external module is built,
it will store the corresponding symbol and version info.
When the module is inserted a check is performed
to ensure that the version for each symbol matches.

This ensures that the interface between kernel
and modules is not broken, and allows more flexibility
than an explicit kernel version check.

(Taken from TI Commit 47e2326bf8cb438e4db9a9d58edf4e931dce4b97)

Change-Id: Ia6acd56d4339565d007a92c409f31137b3e67dac